Refactor structure of simapp test
