+++ app/simulation/app_after_import_test.go
@@ -1,0 +1,136 @@
+package simulation
+
+import (
+	"fmt"
+	"os"
+	"testing"
+
+	abci "github.com/cometbft/cometbft/abci/types"
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/server"
+	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
+	"github.com/cosmos/cosmos-sdk/x/simulation"
+	simcli "github.com/cosmos/cosmos-sdk/x/simulation/client/cli"
+	"github.com/stretchr/testify/require"
+
+	swtrapp "swisstronik/app"
+	"swisstronik/encoding"
+	"swisstronik/tests"
+)
+
+func TestAppSimulationAfterImport(t *testing.T) {
+	config := simcli.NewConfigFromFlags()
+	config.ChainID = SwtrAppChainID
+	config.Commit = true
+	config.NumBlocks = 10 // default value is 500
+
+	db, dir, logger, skip, err := simtestutil.SetupSimulation(
+		config,
+		"leveldb-app-sim",
+		"Simulation",
+		true, // simcli.FlagVerboseValue,
+		true, // simcli.FlagEnabledValue,
+	)
+	if skip {
+		t.Skip("skipping application simulation after import")
+	}
+	require.NoError(t, err, "simulation setup failed")
+
+	defer func() {
+		require.NoError(t, db.Close())
+		require.NoError(t, os.RemoveAll(dir))
+	}()
+
+	appOptions := make(simtestutil.AppOptionsMap, 0)
+	appOptions[flags.FlagHome] = swtrapp.DefaultNodeHome
+	appOptions[server.FlagInvCheckPeriod] = simcli.FlagPeriodValue
+
+	app := swtrapp.New(
+		logger,
+		db, nil, true, map[int64]bool{},
+		swtrapp.DefaultNodeHome, 5,
+		encoding.MakeConfig(swtrapp.ModuleBasics),
+		appOptions,
+		fauxMerkleModeOpt,
+		baseapp.SetChainID(config.ChainID),
+	)
+	require.Equal(t, "swisstronik", app.Name())
+
+	// Run randomized simulation
+	stopEarly, simParams, simErr := simulation.SimulateFromSeed(
+		t,
+		os.Stdout,
+		app.BaseApp,
+		simtestutil.AppStateFn(app.AppCodec(), app.SimulationManager(), simGenesisState()),
+		tests.RandomSimulationEthAccounts,
+		simtestutil.SimulationOperations(app, app.AppCodec(), config),
+		app.ModuleAccountAddrs(),
+		config,
+		app.AppCodec(),
+	)
+
+	// export state and simParams before the simulation error is checked
+	err = simtestutil.CheckExportSimulation(app, config, simParams)
+	require.NoError(t, err)
+	require.NoError(t, simErr)
+
+	if config.Commit {
+		simtestutil.PrintStats(db)
+	}
+
+	if stopEarly {
+		fmt.Println("can't export or import a zero-validator genesis, exiting test...")
+		return
+	}
+
+	fmt.Printf("exporting genesis...\n")
+
+	exported, err := app.ExportAppStateAndValidators(true, []string{}, []string{})
+	require.NoError(t, err)
+
+	fmt.Printf("importing genesis...\n")
+
+	newDB, newDir, _, _, err := simtestutil.SetupSimulation(
+		config,
+		"leveldb-app-sim-2",
+		"Simulation-2",
+		true, // simcli.FlagVerboseValue,
+		true, // simcli.FlagEnabledValue,
+	)
+	require.NoError(t, err, "simulation setup failed")
+
+	defer func() {
+		require.NoError(t, newDB.Close())
+		require.NoError(t, os.RemoveAll(newDir))
+	}()
+
+	newApp := swtrapp.New(
+		logger,
+		newDB, nil, true, map[int64]bool{},
+		swtrapp.DefaultNodeHome, 5,
+		encoding.MakeConfig(swtrapp.ModuleBasics),
+		appOptions,
+		fauxMerkleModeOpt,
+		baseapp.SetChainID(config.ChainID),
+	)
+	require.Equal(t, "swisstronik", newApp.Name())
+
+	newApp.InitChain(abci.RequestInitChain{
+		ChainId:       config.ChainID,
+		AppStateBytes: exported.AppState,
+	})
+
+	_, _, err = simulation.SimulateFromSeed(
+		t,
+		os.Stdout,
+		newApp.BaseApp,
+		simtestutil.AppStateFn(app.AppCodec(), app.SimulationManager(), simGenesisState()),
+		tests.RandomSimulationEthAccounts,
+		simtestutil.SimulationOperations(newApp, newApp.AppCodec(), config),
+		app.ModuleAccountAddrs(),
+		config,
+		app.AppCodec(),
+	)
+	require.NoError(t, err)
+}
