+++ app/simulation/utils.go
@@ -1,0 +1,27 @@
+package simulation
+
+import (
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
+	simcli "github.com/cosmos/cosmos-sdk/x/simulation/client/cli"
+)
+
+// SwtrAppChainID hardcoded chainID for simulation
+const SwtrAppChainID = "swisstronik_1291-1"
+
+// Get flags every time the simulator is run
+func init() {
+	simcli.GetSimulatorFlags()
+}
+
+type StoreKeysPrefixes struct {
+	A        storetypes.StoreKey
+	B        storetypes.StoreKey
+	Prefixes [][]byte
+}
+
+// fauxMerkleModeOpt returns a BaseApp option to use a dbStoreAdapter instead of
+// an IAVLStore for faster simulation speed.
+func fauxMerkleModeOpt(bapp *baseapp.BaseApp) {
+	bapp.SetFauxMerkleMode()
+}
