+++ x/evm/genesis.go
@@ -37,7 +37,7 @@ func InitGenesis(
 	accountKeeper types.AccountKeeper,
 	data types.GenesisState,
 ) []abci.ValidatorUpdate {
-	//k.WithChainID(ctx)
+	k.WithChainID(ctx)
 
 	err := k.SetParams(ctx, data.Params)
 	if err != nil {
